+++ src/main/java/com/dkharchenko_hillel/homework7/services/ProductService.java
@@ -3,10 +3,11 @@ package com.dkharchenko_hillel.homework7.services;
 
 import com.dkharchenko_hillel.homework7.models.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ProductService {
-    void addProduct(String name, Double price, Long shopId);
+    void addProduct(String name, BigDecimal price, Long shopId);
 
     void removeProductById(Long id);
 
@@ -16,5 +17,5 @@ public interface ProductService {
 
     void updateProductNameById(Long id, String name);
 
-    void updateProductPriceById(Long id, Double price);
+    void updateProductPriceById(Long id, BigDecimal price);
 }
